requirements.txt
torch>=2.0.0
torchvision>=0.15.0
scikit-learn>=1.3.0
Pillow>=10.0.0
ultralytics>=8.1.0
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
pytest>=8.0.0
